Refactor out records modules
Using the modules as of Terraform v0.13 has the main issue (for me) of
having to explicitly define the provider requirements used within each
module.

That's fine for reusable modules (that may or may not be published in
the Registry), but in this case, it was mostly a way of modularizing
my resources. I don't mean for those modules to _ever_ be used outside
of this repository.

Which means that, even though I already define the `njalla` provider
requirement at the "root module" (check `config.tf`), I had to pretty
much copy and paste the whole block into each of the submodules as
well.

Then, any changes to the provider version would have required changes
in (as of this commit) 5 different files.

Terraform has a method to inherit provider _configuration_.
Configuration being whatever the provider may require (usually
credentials) inside a `provider name {}` block. However, there doesn't
seem to be any way to indicate to Terraform that the modules will make
use of the same providers (and versions) the root module defines.